requirements.txt
spidev
RPi.GPIO
pyserial
pynmea2